Move validators to own file
